# src/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"_36_Enums.SupplierStatus": {
				"type": "string",
				"enum": [
					"ACTIVE",
					"DELETED",
					"SUSPENDED",
					"BLACKLISTED"
				]
			},
			"DefaultSelection_Prisma._36_SupplierPayload_": {
				"properties": {
					"address": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"contactPerson": {
						"type": "string"
					},
					"companyName": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/_36_Enums.SupplierStatus"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"address",
					"email",
					"phone",
					"contactPerson",
					"companyName",
					"updatedAt",
					"createdAt",
					"status",
					"id"
				],
				"type": "object"
			},
			"Supplier": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_SupplierPayload_",
				"description": "Model Supplier"
			},
			"PaginatedResult_Supplier_": {
				"description": "Shared pagination types used across all services.",
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/Supplier"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"total",
					"page",
					"limit",
					"totalPages"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SupplierStatus": {
				"$ref": "#/components/schemas/_36_Enums.SupplierStatus"
			},
			"SupplierCreationParams": {
				"properties": {
					"companyName": {
						"type": "string"
					},
					"contactPerson": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"taxId": {
						"type": "string"
					},
					"paymentTerms": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/SupplierStatus"
					}
				},
				"required": [
					"companyName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_SupplierCreationParams_": {
				"properties": {
					"companyName": {
						"type": "string"
					},
					"contactPerson": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"taxId": {
						"type": "string"
					},
					"paymentTerms": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/_36_Enums.SupplierStatus"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"_36_Enums.EntityStatus": {
				"type": "string",
				"enum": [
					"ACTIVE",
					"DELETED",
					"SUSPENDED",
					"BLACKLISTED"
				]
			},
			"DefaultSelection_Prisma._36_ProductPayload_": {
				"properties": {
					"minDeliveryDays": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"categoryId": {
						"type": "number",
						"format": "double"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/_36_Enums.EntityStatus"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"minDeliveryDays",
					"description",
					"categoryId",
					"updatedAt",
					"createdAt",
					"status",
					"id",
					"name"
				],
				"type": "object"
			},
			"Product": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_ProductPayload_",
				"description": "Model Product"
			},
			"PaginatedResult_Product_": {
				"description": "Shared pagination types used across all services.",
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/Product"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"total",
					"page",
					"limit",
					"totalPages"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EntityStatus": {
				"$ref": "#/components/schemas/_36_Enums.EntityStatus"
			},
			"ProductCreationParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"categoryId": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"minDeliveryDays": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/EntityStatus"
					}
				},
				"required": [
					"name",
					"categoryId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_ProductCreationParams_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"categoryId": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"minDeliveryDays": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/_36_Enums.EntityStatus"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"DefaultSelection_Prisma._36_ProductCategoryPayload_": {
				"properties": {
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/_36_Enums.EntityStatus"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"updatedAt",
					"createdAt",
					"status",
					"id",
					"name"
				],
				"type": "object"
			},
			"ProductCategory": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_ProductCategoryPayload_",
				"description": "Model ProductCategory"
			},
			"PaginatedResult_ProductCategory_": {
				"description": "Shared pagination types used across all services.",
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/ProductCategory"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"total",
					"page",
					"limit",
					"totalPages"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductCategoryCreationParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/EntityStatus"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_ProductCategoryCreationParams_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/_36_Enums.EntityStatus"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "kdm-procurex-backend",
		"version": "1.0.0",
		"description": "KDM ProcureX Backend",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/suppliers/paginated": {
			"get": {
				"operationId": "GetSuppliersPaginated",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResult_Supplier_"
								}
							}
						}
					}
				},
				"tags": [
					"Suppliers"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/suppliers/{id}": {
			"get": {
				"operationId": "GetSupplier",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Supplier"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Suppliers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateSupplier",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Supplier"
								}
							}
						}
					}
				},
				"tags": [
					"Suppliers"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_SupplierCreationParams_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteSupplier",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Suppliers"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/suppliers": {
			"get": {
				"operationId": "GetAllSuppliers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Supplier"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Suppliers"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateSupplier",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Supplier"
								}
							}
						}
					}
				},
				"tags": [
					"Suppliers"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SupplierCreationParams"
							}
						}
					}
				}
			}
		},
		"/products/paginated": {
			"get": {
				"operationId": "GetProductsPaginated",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResult_Product_"
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/products": {
			"get": {
				"operationId": "GetProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Product"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateProduct",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Product"
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProductCreationParams"
							}
						}
					}
				}
			}
		},
		"/products/{id}": {
			"get": {
				"operationId": "GetProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Product"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Product"
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_ProductCreationParams_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProduct",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Product"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/product-categories/paginated": {
			"get": {
				"operationId": "GetCategoriesPaginated",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResult_ProductCategory_"
								}
							}
						}
					}
				},
				"tags": [
					"ProductCategory"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/product-categories": {
			"get": {
				"operationId": "GetCategories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ProductCategory"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"ProductCategory"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateCategory",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductCategory"
								}
							}
						}
					}
				},
				"tags": [
					"ProductCategory"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProductCategoryCreationParams"
							}
						}
					}
				}
			}
		},
		"/product-categories/{id}": {
			"get": {
				"operationId": "GetCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ProductCategory"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"ProductCategory"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductCategory"
								}
							}
						}
					}
				},
				"tags": [
					"ProductCategory"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_ProductCategoryCreationParams_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCategory",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"ProductCategory"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/v1"
		}
	]
}